@@ -10,8 +10,7 @@ homepage: https://github.com/Idean/Flutter-Neumorphic
 issue_tracker: https://github.com/Idean/Flutter-Neumorphic/issues
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.13.18"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -23,6 +22,6 @@ dev_dependencies:
 
 flutter:
   fonts:
-   - family: NeumorphicIcons
-     fonts:
-      - asset: fonts/NeumorphicIcons.ttf
+    - family: NeumorphicIcons
+      fonts:
+        - asset: fonts/NeumorphicIcons.ttf
